Put types with their functions in separate files and make colors work on Windows
