+++ go.mod
@@ -3,9 +3,9 @@ module github.com/herver/go-gvacli
 go 1.12
 
 require (
-	github.com/fatih/color v1.7.0
-	github.com/mattn/go-colorable v0.1.2 // indirect
 	github.com/mattn/go-runewidth v0.0.4 // indirect
 	github.com/olekukonko/tablewriter v0.0.1
 	github.com/spf13/pflag v1.0.3
+	github.com/stretchr/testify v1.4.0 // indirect
+	gopkg.in/gookit/color.v1 v1.1.6
 )
